packages/backend/pyproject.toml
[project]
name = "verbatim-backend"
version = "0.1.0"
description = "Verbatim Studio Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
ml = [
    "whisperx>=3.1.0",
    "torch>=2.0.0",
]
export = [
    "python-docx>=1.0.0",
    "reportlab>=4.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.setuptools.packages.find]
include = ["api*", "core*", "persistence*", "services*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
